requirements.txt
autogen
openai 
python-dotenv
flaml[automl]
matplotlib
numpy